requirements.txt
Django==3.2.15
openpyxl==3.0.10
pytz==2022.2.1
requests==2.28.1
pandas==1.3.5
beautifulsoup4==4.11.1
python-telegram-bot